requirements.txt
pymongo
python-dateutil
pytz
crawl4ai
dnspython